feat: add database types and sample data for vendors and listings; update search engine dimensions and text viewer store
