+++ src/data/listings.js
@@ -1,0 +1,99 @@
+// Sample listings for the marketplace
+export const LISTINGS = [
+  {
+    id: "NSE-1042",
+    title: "Network Penetration Suite 2.0",
+    description:
+      "Complete toolkit for bypassing network security protocols. Tested on major enterprise systems with 94% success rate. Includes zero-day exploits and anti-detection modules.",
+    price: 5999,
+    ratings: 4.8,
+    transactions: 124,
+    category: "Software",
+    vendor: "CobraSystems",
+    verified: true,
+  },
+  {
+    id: "MED-7734",
+    title: "Medical License Generator",
+    description:
+      "Premium medical credential creation package. Includes templates for 14 specialties with embedded authentication markers that pass standard verification checks.",
+    price: 4500,
+    ratings: 4.7,
+    transactions: 87,
+    category: "Documents",
+    vendor: "GhostDoc",
+    verified: true,
+  },
+  {
+    id: "IND-8821",
+    title: "Industrial Control System Backdoor",
+    description:
+      "Specialized software for accessing industrial control systems. Created for SCADA networks with multi-layer security. Leaves minimal log footprint and includes remote activation modules.",
+    price: 12000,
+    ratings: 4.9,
+    transactions: 36,
+    category: "Software",
+    vendor: "Prometheus_X",
+    verified: true,
+  },
+  {
+    id: "NSE-1055",
+    title: "Corporate VPN Takeover Kit",
+    description:
+      "Advanced toolkit for compromising enterprise VPN systems. Targets major VPN providers with proprietary exploitation methods. Includes step-by-step documentation.",
+    price: 8500,
+    ratings: 4.6,
+    transactions: 53,
+    category: "Software",
+    vendor: "CobraSystems",
+    verified: true,
+  },
+  {
+    id: "MED-7736",
+    title: "Prescription Authorization Package",
+    description:
+      "Complete system for generating authorized prescriptions that pass pharmacy verification systems. Updated monthly with current authorization codes.",
+    price: 3200,
+    ratings: 4.5,
+    transactions: 142,
+    category: "Documents",
+    vendor: "GhostDoc",
+    verified: true,
+  },
+  {
+    id: "IND-8823",
+    title: "Infrastructure Blueprint Collection",
+    description:
+      "Classified technical documentation for critical infrastructure systems. Includes detailed schematics and security protocols for power grid implementations.",
+    price: 15000,
+    ratings: 4.8,
+    transactions: 21,
+    category: "Documents",
+    vendor: "Prometheus_X",
+    verified: true,
+  },
+  {
+    id: "ID-2240",
+    title: "Premium Identity Package",
+    description:
+      "Complete identity set with full background verification. Includes digital footprint history and supporting documentation that passes standard verification checks.",
+    price: 7500,
+    ratings: 4.7,
+    transactions: 68,
+    category: "Identity",
+    vendor: "MirrorMask",
+    verified: true,
+  },
+  {
+    id: "DAT-5578",
+    title: "Corporate Data Bundle - Finance Sector",
+    description:
+      "Comprehensive data collection from major financial institutions. Includes customer records, transaction histories, and internal communications from Q1-Q3 2023.",
+    price: 20000,
+    ratings: 4.9,
+    transactions: 17,
+    category: "Data",
+    vendor: "QuantumHarvest",
+    verified: true,
+  },
+];
