+++ src/data/vendors.js
@@ -1,0 +1,52 @@
+export const VENDORS = {
+  CobraSystems: {
+    id: "CobraSystems",
+    realName: "Alex Karimov",
+    specialty: "Network Security Exploits",
+    status: "Missing",
+    lastSeen: "March 12, 2023",
+    clue: "Small toy snake on circuit board found at scene",
+    profileQuote: "Security holes are just opportunities waiting to be found.",
+    accessUnlocked: false,
+  },
+  GhostDoc: {
+    id: "GhostDoc",
+    realName: "Dr. Leyla Mahmudova",
+    specialty: "Medical Credentials and Prescriptions",
+    status: "Missing",
+    lastSeen: "March 17, 2023",
+    clue: "Antique medical caduceus wrapped in gauze",
+    profileQuote: "Your medical identity is just a prescription away.",
+    accessUnlocked: false,
+  },
+  Prometheus_X: {
+    id: "Prometheus_X",
+    realName: "Ibrahim Nasirov",
+    specialty: "Industrial Sabotage Software",
+    status: "Missing",
+    lastSeen: "March 24, 2023",
+    clue: "Small metal lighter with Greek lettering",
+    profileQuote: "Bringing forbidden knowledge to mankind.",
+    accessUnlocked: false,
+  },
+  MirrorMask: {
+    id: "MirrorMask",
+    realName: "Unknown",
+    specialty: "Identity Theft and Impersonation",
+    status: "Active",
+    lastSeen: "Recent Activity",
+    clue: "Unknown",
+    profileQuote: "Everyone wears a mask, I just collect them.",
+    accessUnlocked: false,
+  },
+  QuantumHarvest: {
+    id: "QuantumHarvest",
+    realName: "Unknown",
+    specialty: "Data Harvesting and Information Sales",
+    status: "Active",
+    lastSeen: "Recent Activity",
+    clue: "Unknown",
+    profileQuote: "Your data is our most valuable crop.",
+    accessUnlocked: false,
+  },
+};
